requirements.txt
PySide6>=6.5.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
pillow>=10.0.0
# Added for global hotkey and active window detection
pynput>=1.7.0
PyGetWindow>=0.0.9
psutil>=5.9.0
pywin32>=306; sys_platform == 'win32' # Windows specific API
google-auth-httplib2
requests
# Added for AES encryption
cryptography